pyproject.toml
[tool.poetry]
name = "hinglish-service"
version = "0.1.0"
description = "Stateless Hinglish transcription service using Gemini API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = ">=1.1.0,<2.0.0"
python-multipart = "^0.0.17"
google-genai = "^1.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"